+++ vtrace/tests/testwritemem.py
@@ -13,7 +13,7 @@ class VtraceWritememTest(vt_tests.VtraceProcessTest):
         self.runProcess()
 
         addrstr = self.proc.stdout.readline()
-        addr = long(addrstr, 16)
+        addr = int(addrstr, 16)
 
         testbuf = binascii.hexlify(os.urandom(10))
